+++ src/main/kotlin/com/example/running/domain/running/controller/dto/UpdateRecordRequest.kt
@@ -2,7 +2,7 @@ package com.example.running.domain.running.controller.dto
 
 class UpdateRecordRequest(
     val shoeId: Long? = null,
-    val distance: Long? = null,
+    val distance: Int? = null,
     val durationSec: Long? = null,
     val cadence: Short? = null,
     val heartRate: Short? = null,
